move documentation and feedback to hamburger menu
